requirements.txt
numpy==1.15.2
Pillow==5.3.0
pytesseract==0.2.4
